Extract client commands handling to separate class + minor fixes
