+++ client/CMakeLists.txt
@@ -81,6 +81,7 @@ set(client_SRCS
 		NetPacksClient.cpp
 		NetPacksLobbyClient.cpp
 		SDLRWwrapper.cpp
+        ClientCommandManager.cpp
 )
 
 set(client_HEADERS
@@ -165,6 +166,7 @@ set(client_HEADERS
 		mapHandler.h
 		resource.h
 		SDLRWwrapper.h
+        ClientCommandManager.h
 )
 
 if(APPLE_IOS)
